# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinPilot",
    "description": "Authenticated financial workspaces with deterministic calculations.",
    "version": "2.0"
  },
  "paths": {
...
    "/api/ask": {
      "post": {
        "tags": [
          "Assistant"
        ],
        "summary": "Ask the assistant",
        "description": "Answers a question or request inside a saved conversation. Numbers come from deterministic calculations, and model wording carries figure tokens that the server fills in and checks. Answers link the household records behind them, name the result field behind each figure and carry an evidence confidence level. Unrelated requests get a fixed reply without a model call. Requested changes come back as proposals that need confirmation. Rate limited per user and by assistant capacity.",
        "operationId": "ask_api_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "403": {
            "description": "Cross-site request, failed CSRF check, or the role or workspace type does not allow this action."
          },
          "404": {
            "description": "The record does not exist in this workspace."
          },
          "409": {
            "description": "The workspace revision changed, or the record's current state does not allow this action."
          },
          "422": {
            "description": "The request failed validation."
          },
          "429": {
            "description": "A rate limit or capacity limit was reached; retry shortly."
          }
        }
      }
    },
    "/api/ask/history": {
      "get": {
        "tags": [
          "Assistant"
        ],
        "summary": "List saved answers",
        "description": "Lists the signed-in user's recent saved answers in this workspace.",
        "operationId": "history_api_ask_history_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          },
          "422": {
            "description": "The request failed validation."
          }
        }
      }
    },
    "/api/ask/suggestions": {
      "get": {
        "tags": [
          "Assistant"
        ],
        "summary": "Get example questions",
        "description": "Returns example questions for the assistant.",
        "operationId": "suggestions_api_ask_suggestions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Not signed in, the session expired, or the MCP read token is missing, expired or revoked."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AskIn": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 3000,
            "minLength": 1,
            "title": "Question"
          },
          "untrusted_context": {
            "type": "string",
            "maxLength": 8000,
            "title": "Untrusted Context",
            "default": ""
          },
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ViewingContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "tool_choice": {
            "$ref": "#/components/schemas/ToolChoice",
            "default": "router"
          },
          "include_evidence": {
            "type": "boolean",
            "title": "Include Evidence",
            "default": true
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 40
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "document_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "maxItems": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Ids"
          },
          "workflow_mode": {
            "type": "string",
            "enum": [
              "auto",
              "model"
            ],
            "title": "Workflow Mode",
            "default": "auto"
          },
          "workflow_input": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/OperationIn"
                },
                "type": "array",
                "maxItems": 4,
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Input"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AskIn"
      },
...
      "OperationIn": {
        "properties": {
          "capability": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Capability"
          },
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "capability"
        ],
        "title": "OperationIn"
      },
...
      "ToolChoice": {
        "type": "string",
        "enum": [
          "router",
          "model",
          "auto"
        ],
        "title": "ToolChoice"
      },
...
      "ViewingContext": {
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "page": {
            "type": "string",
            "maxLength": 40,
            "title": "Page",
            "default": "overview"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ViewingContext"
      }
    }
  }
}